fix/refactor select_hud
the cursor is still broken, but select_hud now works and is a bit
improved.  this commit removes the middleman bullshit garden was doing
and lets the select hud handle its own affairs.  it also adds a fancy
delayed fade-out effect when you unhighlight the selected entity.

this commit also adds a dinky minute hand to clock_hud instead of the
pendulum.  i think i like this better but it needs proper assets to
really work (the minute hand is the same as the hour hand but reddish)