fix move action
updates it to use NavigationAgent2D on the monster.  this works
because we have navigation data on the garden's tilemap, which is
automagically baked into the default world navigation map/mesh.
unfortunately it plots paths up against the edges of navigable tiles,
meaning that if there were collision the monster would get stuck on
corners (and generally hug walls in an unrealistic way).

previously we solved this by messing with each point in the path
generated by the old Navigation2D that was part of the garden (file is
kept in `_scratch/navigation/nav.gd`) to find the center of the closest
navigable tile and move it there instead.  this worked as a one-size
solution but was probably not very performant.

the "right" way to solve this, presumably, is to custom-bake a navmesh
with a certain amount of padding around all walls - shrink the navigable
area.  this has some nice benefits:

- objects don't have to block the entire tile they're on, so some
  adjacent objects could have navigable gaps between
- can bake multiple navmeshes for different sizes of monster and choose
  the appropriate one when pathing - allows smaller monsters to fit
  where larger ones can't

alterately, there's this hack:
https://old.reddit.com/r/godot/comments/x7s4z1/guide_for_how_to_fix_navigationagents_getting/

might be ok for a dirty solution but the commenter there explains why
it's not optimal.  we should just look into home baking.