fix attempts: PerfectScale & monster movement
perfect_scale:

adjusted stuff to use Vector2i instead of Vector2 (to match the new
window/viewport sizes).  the scene tree also no longer has a
`screen_resized` event, we have to listen to one on the Window at the
root of the sceen tree (`get_tree().get_root()`, or its equivalent
according to the docs, `get_window()`).

previously, if the window was fullscreen we would add a gutter, and if
not we would attempt to resize it.  however, it turns out that window
mode is a bad proxy for whether the window is resizable - in i3wm the
window mode is still WINDOWED whether it's floating (resizable) or tiled
(not).  in the latter case, PerfectScale repeatedly attempts to resize
the viewport, which triggers a `size_changed` signal even though the
size didn't actually change, so it thrashes endlessly.  this is still
busted and needs to be fixed, but at least it works well with resizable
windows now.

my plan was disable calling `update_screen` on `size_changed` for long
enough to stop the infinite loop, if the size didn't actually change.
this is complicated by the need to debounce, so `update_screen` only
gets called at most every x seconds (resizing windows by dragging would
fire a lot of events in a short time). i tried debouncing with a simple
delta counter instead of a timer class, but _process and _physics_process
don't actually seem to tick on `perfect_scale.gd`.

monster movement:

`move_and_slide` was super slow at first and then sped up gradually,
even though the monster's velocity vector stayed a consistent length.
it's really noticeable when clicking repeatedly (each click creates a
new move action for the test monster), so i figured it might be because
there were multiple move actions operating on the same physics frame.
i added some logic in set_current_action to end the current action
before setting the new one, and it seemed to help a little bit, but
i still notice bunny speeding up on repeated clicks.  it may not be fast
enough to keep up with input - maybe this needs its own debouncing.