sorta fix perfect_scale
lots of comments on the perfect_scale file.  we cannot make it work for
gutters because the viewport is the window, and always gets re-stretched
to the window size.

the solution is obviously to add a sub-viewport for the actual game, but
i was stubborn enough to keep going down this dead end for like 2 days
:(