sub-viewport success
adds a sub-viewport under `game.tscn` to handle pixel-perfect scaling,
and moves a stripped-down version of PerfectScale in there (we no longer
need an autoload, though i guess scenes run on their own will no longer
be PerfectScaled - not sure if that matters though).

the new code doesn't bother resizing the window, it only scales the
sub-viewport as large as it can go and then adds the gutter.

there's still a bug where we sometimes come up with a scaled_size
greater than the current window size.  this was ok when we were resizing
the window (though not ideal - what if the window was already as big as
the screen?), but now it needs to be handled or the scaling is no longer
perfect.