inventory fixup
- fix some visual stuff
- remove more config from `constants.gd`

note: updating item_info still isn't working.  it's listening to a
dispatcher signal `item_selected`, but nobody fires that signal, and the
obvious place to fire it (`update_current_item` in `item_grid`) is
called from the signal handler (`update_current_item` in `inventory`).

i'm not sure how this was set up before and how it ended up broken, but
it seems to need a rethink ¯\_(ツ)_/¯