reset command & ui input fixes
moves the console out of the ui stack to its own CanvasLayer.
- this allows us to manipulate the ui stack while the console is open
  without worrying about keeping the console on top.
- it also lets us keep the console alive while it's closed (hidden), so
  that output and past commands are preserved.
- for now, opening the console doesn't pause the game anymore.
  i think i like this better, but we can always change it back later -
  we will just need to conditionally unpause on closing the console (should
  stay paused if the ui stack is nonempty).

moves ui input signal handling into ui.gd, out of the dispatcher.  this
surfaced a bug: signal handling was broken in children of the game node
because its process mode was set to "while paused".  (ui input still
worked because Dispatcher, an autoload, sits outside the scene tree.)

with that fixed, the garden handles input as expected again, so we can
once again click to direct monsters.  since we can spawn more monsters
now, this will queue moves for _all_ the mons in the garden.