rename traits to attributes
originally, monster "traits" were stateful and "attributes" were
computed from traits.  however, we want to use a trait pattern to
implement behaviors for objects and items, and picking a different name
for that would be confusing.

this commit renames every instance of "trait" (referring to monster
traits) to "attribute".  it also adds new computed properties to the
Attribute class for the values formerly known as "attributes", named in
all caps to distinguish them from regular attributes (INT, VIT, etc).