adjust monster sizes & fix offsets
fixes a bug where monster sprite offsets weren't getting updated when
the texture didn't change (eg, when left and right facings for an anim
use same texture flipped horizontally).

note: by default animations set the value on every frame.
UPDATE_DISCRETE mitigates this but it may fail to set the value when we
change facing direction in the middle of an animation, since we seek
(skip) to the new animation position.