add Entity class
create a superclass of Monster that will be shared with items and
objects/sessiles.  this allows us to avoid duplicating initialization &
serialization code and common child nodes like `sprite.gd`.

Entity inherits RigidBody2D, so all of our entity subclasses must as
well.  this is fine since we just ported monster to RigidBody, we were
going to use it for items anyway, and it will work for objects as long
as we set the frozen flag so they can't be moved (but can still simulate
physics on collisions like bouncing).