add back steering, fix cursor v-offset
cleans up `move.gd` and reenables the steering code.

also cleans up the cursor code a bit and fixes a bug that was preventing
the cursor, when stuck to a monster, from picking the right height based
on the monster's shape (we were checking the monster's CollisionShape2D,
not the shape contained in the CollisionShape's `shape` property).

the cursor will now only stick to entities (subclasses of the new Entity
class).

plus more movement tuning:

- increase default linear damp to 10 (0 velocity in 1/10 of a second)
  - add this directly to desired velocity in `move.gd`
  - throw this calculation in a static funcion so we can reference it
    elsewhere (debug raycasts in `monster.gd`)
- make milotic way lighter (and a little smaller)