(wip) grab + orientation
first stab at making monsters (attempt to) hold whatever they are
grabbing in front of them.

we add a Joint2D to the monster to represent its grabbing organ, with
`node_a` permanently set to the parent monster.  to grab something
else, we set `node_b` to its nodepath.

using a joint is nice because it automatically gives us the physics
behavior we want: mons will move slower when pulling objects (same force
pushing more mass), and bigger/heavier mons will tow smaller ones around
regardless of who is the grabber and who is the grabbed.

i could find two ways to change the orientation of the grabbed entity
with respect to the grabber:

1. with rotation off: using a spring joint, the grabbed can rotate
   around the collision point.  there's no way to change the center of
   rotation, so it will be further away from the grabber on one side
   than on the other.  also, the grabbed floats a bit away from the
   grabber even when the spring joint's length is 0.

2. turn rotation on for entities.  this allows the two attached bodies
   to swing around each other the way we want.  coupled with movement
   logic that applies the force to the grabbed object so as to keep it
   in front, the physics work perfectly.

unfortunately we do NOT want monster sprites (or other children, eg
debug raycasts) to rotate.  we can keep them oriented correctly by
cancelling out the parent's rotation on each physics frame, but this
breaks offsets - eg, when the entity is rotated 180 degrees, the bottom
of the sprite is still aligned with the "bottom" of the collision shape,
but the bottom of the collision shape is now at the top.