fix rotation offset issue
fixes the bug with sprite offsets when allowing rotation on the parent
RigidBody2D and compensating for it on the child sprite node.  turns out
the `offset` prop on the sprite was fine, only the `position` we set was
affected by the parent's rotation.

this commit adds the parent entity's `size` to the sprite offset instead.
for performance we save it in the sprite on load so we don't have to
keep referencing `get_parent().data.size`.  that's one more potential
bug to fix when we eventually need to support changing a monster's type
at runtime, but we'll cross that bridge when we get there ¯\_(ツ)_/¯

i'm impressed with how well this works, tho there are still some bugs:

- grabbing seems to mess with the garden's `y_sort_enabled` - grabbed
  entities at lower y-positions (closer to the top) are sometimes
  layered on top of their grabbers instead of behind

- with lots of monsters all grabbing each other, sometimes the physics
  engine freaks out and flings them away at super speed.
  this probably has something to do with multiple milotics grabbing a
  smaller monster - milotic (90kg) generates an order of magnitude more
  force to move itself than bunny (3kg) does.

also, here is a useful reference i forgot to add to the last commit msg:
https://old.reddit.com/r/godot/comments/ekwwbp/make_child_node_ignore_rotations_of_the_parent/