fixes & cleanup
- limit entities' linear velocity to 100 in `integrate_forces`
- detach grabs if the bodies get too far apart

this mitigates a lot of the freakouts but it still happens to a lesser
extent (looks like 1 milotic + multiple bunnies is a good test case).

- maybe fix nan vector bug

this is/was a bug with serializing vectors: if the vector coords are
too large they get serialized as `nan` (not `"nan"`), which breaks json
parsing on next load.  as of `serialize_value` they are NOT considered
`nan` yet, so we attempt to JSON.stringify them in there and leave them
out of serialization if the values are too big for JSON.  i haven't
actually tested this but hopefully it works lol