homebrew translation support
since we are adding a bunch of new ui text with monster details, now
seems like a good time to add some kind of translation support so it's
less of a headache in the future.

this commit adds `trans_text.gd` as a repository for translatable text
strings.  the idea is to use the constants in this file plus the `tr`
method for all translatable text (eg, `tr(T.MONSTERS)`).

this includes text that goes on labels and buttons, instead of relying
on godot's automatic translation if the label/button text is a trans-
lation key.  this means we should always set the text's value in code
(ie, in `_ready`) and use some placeholder text in the editor.
this has a few advantages:

1. consistency with variable/dynamic translatable text which _must_ be
   set in code and thus must use `tr`
2. ability to call functions like `capitalize` and `to_upper` on the
   result of the translation
3. the advantages of using constants for translatable text in general:
   decoupling and compile-time guarding against missing keys

this commit also switches to a 9pt version of the font (generated by
setting the em size to 9/16 of the original 1024), so that if we need to
use fallback glyphs from some other font they will be approximately
similar in size to the regular glyphs.