fancy tabs
moves main menu tabs to `tabs.gd`, which makes them look more like
bookmarks by keeping the current tab in the center and moving the others
to the left or right.  this has the added benefit of cleaning up the
main menu code a little bit.

also in main menu cleanup, removes the `content` node which (i think)
was there to center stuff but wasn't actually necessary, and `arrows2`.

TODO: flip z-order of tabs in $next. setting z-indices is a crapshoot
because they end up either over the menu or under the background, so
flipping the direction and reparenting them in reverse order is
probably the way to go.