menu stuff
reworks menu_chapter to be a script instead of a scene, and changes our
empty menu chapters (town map, calendar, etc) back to MenuSections.

chapters now build their index on init, and main_menu maintains a dict
of chapter keys to MenuChapter instances that are initialized onready,
meaning opening the menu will call build_index for every chapter.  this
may not be ideal once more chapters have more stuff, but it's convenient
for now and easy enough to fix later.

monster_list is now a PagedList, which means monsters in the list are
organized by row (left-right then top-down) instead of by column like
they were previously.  this doesn't actually matter but we can fix it if
we want by reordering them in `load_items`.

this commit also updates `main_menu.open` to take in a path instead of
just a chapter key, and load the appropriate section, so we can now
successfully navigate between monster list and monster info.

TODO:
- add arrows to nav between sections (when permitted)
- fix bug with page arrows sometimes not being clickable