inventory menu refactoring
item_grid evolved into PagedList!!

also deletes the `items` & `objects` subclasses of `inventory.tscn` and
replaces them with a config const.  this was inelegant and caused a bug
because those subclasses were calling `initialize` in _ready, while the
new menu_chapter implementation calls it on instantiate, meaning it was
being called twice.