kill Dispatcher.emit
this function was from before godot 4; it called `emit_signal` with
packed args and fired a log line.

Dispatcher.signal.emit is a nicer API since it has autocompletion and
compile-time errors for missing signals (i think), and the logging and
argument packing  weren't really necessary.

this commit removes Dispatcher.emit and replaces all usage of it with
Dispatcher.signal.emit, except where we have to use emit_signal because
the signal name is being dynamic (eg Clock).