set up entity sprites with a function
on monster init, we build animations for each facing of each animation
key in the monster morph's data definition.  the `_init_facing` function
(now renamed to `create_animation_for_facing`) in `anim.gd` previously
created a TYPE_VALUE track for each sprite property to be updated with a
single keyframe at 0.0 seconds.

this commit moves that logic to a new `update_texture` function in
`sprite.gd`, and updates `create_animation_for_facing` to add a single
TYPE_METHOD track that calls `update_texture` with the relevant
anim_info for the facing.

this will eventually allow us to build monster portraits that can render
a clipped sprite with the correct texture info for a given animation key
and facing.