fancy select_hud portraits
all we actually need to align the sprite are texture, hframes, and flip.
we use a panel container that extends a few pixels past the top of the
frame to mask the sprite.  if the sprite is larger than the container,
we top-right align it (assuming it's always using front-right facing).
if it's smaller, we center it.

only the select hud will have round portraits; portraits on the main
menu will use item boxes with rounded corners.  we should be able to use
the same component for these if we make it into a tool and export all
the necessary configuration.