use portrait in monster details
this commit turns `portrait.tscn` into a tool which exports all the config
we need to use it with different kinds of containers, and updates
`name_card` in `monster_details` to use it.

TODO: handle x and y axis separately instead of doing one size
comparison, which breaks for monsters wider than the container but not
as tall (eg sleeping bunny).