use portrait in monster list
adds portrait.tscn to the third and final place (for now) where it needs
to be, the monster list in the main menu.

this updates the `update` method on portrait to take an entity and fetch
the sprite info from it, instead of having to do that beforehand in each
of its callers.  it also adds two new Entity helper methods (implemented
for Item and Monster), `get_sprite_info` and `get_display_name`.

it also updates portrait to be a NinePatchRect rather than a simple
TextureRect - it can still handle regular textures while also being more
flexible.