(rough) return errors from validator
updates the validator to return an error digest instead of logging.
this makes all the functions return a { valid, errors } dict instead of
an integer, updates `report` to return a { valid, error } dict instead
of logging, and replaces each `result &= sub_result` (where both were 1
or 0) with `append_sub(result, sub_result)`, which &='s their respective
`valid` properties instead and appends the sub_result to `result.errors`
if it is invalid.

this is a first pass; it's still pretty messy.  our actual use cases
aren't that robust so i'll be adding tests and refining the output
format that way (this change was mostly to make validator tesable in
the first place).

note that this output format does NOT match any of the official
json-schema output formats and likely never will; those require keyword
locations (ie schema breadcrumbs), both relative and absolute, and we
don't track those.