make validator not suck
that's more like it :)

this replaces  `schema_path` (s) & `absolute_schema_path` (sa) in all
the validate methods with a reference to a `schema_ctx` object.  this
can have one or both of the two path properties; unnecessary ones are
omitted (eg, absolute path is redundant if the schema doesn't resolve
a ref).  this is managed through the `schema_context` function, which
is the source of truth for the schema path property keys so that they
can be easily renamed.

`report` and `validate_instance` (the two functions that generate new
result objects) now optionally take a schema context and a string for
the next part of the path.

`result` (renamed `new_result` to avoid conflicts) merges the context
into the result object.

this commit also (finally) adds a few test cases for the validator :)

and, now that we've proved that $refs and sibling properties interact
the way we want, we can split out a general entity schema with shared
properties into its own file, and reference it from all the different
entity schemas.

and, now that we've proved that $refs and sibling properties interact
the way we want (combined as though with `allOf`), we can move shared
entity properties into a generic `entity.schema` file referenced from
the monster, item and object schemas.  this will both DRY our schemas
and make it easier to maintain parity between schemas and entity code
(ensuring that properties required by the implementation are required
in the schema as well).