validator gddoc comments
turns out you actually can populate the generated doc pages for custom
classes - these use ## instead of # and bbcode for formatting.  just for
fun, this commit reformats all the method comments in validator.gd as
gddoc comments.

it's neat to see the result in a generated doc just like a real godot
class, but unfortunately the bbcode formatting makes the comments less
readable in the actual code, which is where we'll be looking at them
most of the time, so it's not really worth the effort.  also, method
comment output doesn't currently support paragraph breaks or lists, so
in many cases they will be less readable than code comments anyway.

in other news, this commit adds a couple of new tests, simplfies error
messages (`invalid value` vs of `does not have an acceptable value`),
and fixes a bug in `uniqueItems` where we were calling `Array.push`
instead of `Array.push_back`.

this is the last validator commit i swear!!