handle defocusing in PagedList
fixes a bug where clicking outside of a selected item in PagedList would
unfocus it (godot thing, visual change & lets you interact with items
using the action button) but not deselect it (PagedList thing, supports
pagination / context-sensitive movement input).

note that there is a third way for a PagedList item to lose focus:
pushing another control to the ui stack.  in theory multiple PagedLists
could be rendered at the same time, each with their own selection, but
only one control can be focused at a time.  this is still an outstanding
issue.  (repro on the title screen: open the save list, then hit
ui_cancel to go back to the button menu. the "load game" button will be
selected but not focused.)