reset default ui inputs
turns out that default input actions like `ui_menu` are used internally
by ui nodes, which is why they can't be deleted.  in particular,
`ui_menu` is used to summon a context menu in editable fields like
LineEdit.

generally, components that handle built-in actions do so before the
actions get propagated to gdscript input handling.  eg, when `ui_menu`
was bound to the M key, pressing `m` in a LineEdit would bring up the
context menu _and_ thekeypress would be passed to `_unhandled_input`.

the solution is to avoid repurposing built-in ui actions for other stuff
(eg, opening the in-game menu); we want new actions for that.  this
commit adds a new `menu` action in lieu of overloading `ui_menu`, and
drops the `ui_` prefix from our custom actions (`console` and the
various menu chapters) to distinguish them from built-in actions, which
all start with `ui_`.

i haven't tested joypad control yet and this change likely breaks them.
we will probably want to bind joypad keys to `ui_accept` and `ui_cancel`
(again).