fix new game name input
- autofocus the name input & make the caret blinky
- update LineEdit/TextEdit colors
- actually fix the `ui_open` signal on the new game button.
  we need to _not_ reopen the title screen after closing the name input,
  or we won't be able to see the garden.

TODO: it is time to refactor the ui stack.

the title page can be dismissed via `ui_cancel`, leaving nothing but a
blank screen.  we need to distinguish between pages that should be
treated as modals (closable with `ui_cancel`) and pages that shouldn't,
while still allowing them to be closed programmatically (eg title screen
closing on transition to garden).

passing the stack index to ui_open is pretty much useless.  i think we
can get away with just a `push` method and a `replace` method (keeping
the option to save and restore the replaced ui stack because why not).