action prerequisites and grabbing
adds the `prereq` property to Action, which stores another Action.
if this is nonnull, `tick` and `proc` delegate to the prereq instead
of calling the action's own internal `_tick` and `_proc` methods.
(this works recursively, as a prerequisite action might have its own
prereq that it's delegating to, and so on.)  once the prerequisite
action exits, we clear the `prereq` variable and execution proceeds
as normal.

this allows us to check that prerequisites are still met at any point
during execution of an action, and queue up a new prereq action if
necessary.