energy capacity & calc tweaks
adds energy capacity as a multiple of BMR (which is in kcal/day, so the
idea is that we can store x days worth of energy (currently 2)).

on testing, it turns out that the energy cost of movement (~0.2 kcal on
a typical move action lasting maybe a dozen ticks, with the 16 pixels
per meter adjustment) is a tiny fraction of BMR upkeep (~0.6 _per tick_
for bunny).

it also turns out that this is (fairly) realistic!  BMR apparently
accounts for up to 75% of calories burned in humans, and presumably even
more for rabbits since they (also presumably) have more lean muscle
mass.  it seems that even while exercising, the calories you burn from
that exercise are unlikely to exceed the calories you were already
burning just by existing.

well, while realism is cool, this isn't super fun from a gameplay
perspective - in particular, energy usage and target energy are the
lever we use to make high-pep monsters display a higher preference for
activity than low-pep ones.  and more generally, in terms of gameplay
semantics, energy is more meaningful in the context of activity than in
the context of upkeep.

so basically we need to tweak the numbers on movement energy costs to
make them more significant.  this is also a good opportunity to deter-
mine a range of valid mass-specific BMR values that will ensure that
energy attrition rates always make sense relative to activity costs --
in other words, use activity costs as a baseline -- while still allowing
for meaningful differences in BMR, eg hummingbirds vs snakes.