eating!
implements bare-bones behavior for the `eat` action: on tick, if the
monster is grabbing the target, add its energy sources to the monster
and its mass to the monster's belly, then free the target.

to make this work we had to update `choose_action` to await the next
physics frame so `perception.get_overlapping_bodies` will return an up-
to-date list, otherwise the freed edible would still be in there.