mostly fix approach overshooting
turns out the root of the problem was that `is_navigation_finished`
doesn't seem to respect `target_desired_distance`.  this commit swaps it
out for `is_target_reached`, which seems to do the trick, and updates
our defaults for `target_distance`.

we also update `require_in_range` for grabbing to check that the monster
is colliding with the target, which removes some guesswork about how far
away the two actually have to be (though `monster.size + target.size`,
which is our default grab range, should also be fine, so maybe we didn't
need to change this.)

finally, we also add a new `require_target` requirement to eat, grab and
approach, which fails the action if the target no longer exists.