pure energy
the joke is that this commit makes the sleep action restore 0.5% of the
monster's energy per tick (modified by vigor), rather than doing nothing
on its own ("pure" as in it comes out of nowhere, rather than coming
from food, har har)

in addition to this we reimplement metabolism / catabolism according to
the latest plan in biology3.md:

- energy sources decay at a rate of (currently) 5% per tick, which
  ensures they stay at reasonable levels
- we convert the decayed energy into a multiplier on a "standard" energy
  recovery rate of 0.5% capacity per tick (same as sleep recovery),
  using a few magic numbers which may be tweaked later
- this multiplier ranges from -0.2 (0.1% of energy capacity lost) to 1.5
  (0.75% of energy capacity gained); thus we can either gain energy per
  tick (in addition to sleep gain if the monster is sleeping) or lose it
  if our energy sources are depleted (but at a rate much slower than
  attrition previously, and also slower than sleep gain).

because sleep always restores more energy than the maximum energy loss
from metabolism, monsters will always be able to sleep to regain enough
energy to act even if they are starving, preventing death loops.
because the minimum multiplier is less than 0, monsters cannot stay at
energy equilibrium just by spamming idle - eventually they will have to
sleep to restore energy.

the new code does away with BMR for now (though i haven't removed it
from the schema).  it turns out that more complexity is a bad thing if
it isn't actually needed, and to that end i'd like to experiment with
modifying mean appetite per species as a proxy for mass-specific BMR.

this commit also updates the select hud to use MultiTextureProgressBars
for belly and energy instead of printing a percentage.  it looks nicer
_and_ it's easier to parse!