exhaustively document map.gd
why did i spend... 3? 4 hours writing this? who knows ¯\_(ツ)_/¯

i mean, i guess it's because the map stuff is _extremely_ self-contained
and i wanted to put as fine a bow on it as possible before moving on to
something new.  my usual method for getting back into this is to review
whatever i last did before leaving off, just not usually to this extent.
but i've been in a pretty serious creative slump lately, ie not quite up
to starting a whole new feature, so i guess i ended up picking at this
file like a scab as a way of doing something that felt like progress
without actually having to think much.

and i mean, i guess more documentation isn't a bad thing, even if it's
not the most efficient use of time.  this wang shit is complicated!

but anyway, enough of that! we're warmed up, let's boogie! next up:
CURSOR STATES HELL YEHA