first pass at garden interaction - highlighting
this commit adds some of the scaffolding for interaction states to the
garden.  it moves highlighting behavior out of the cursor and into the
garden itself, and adds a fancy outline shader to highlighted entities.

to make this outline shader work, we have to wrap all entity sprites in
a CanvasGroup, since fragment shaders on sprites can't extend past the
bounds of the sprite, and our sprites are generally not padded.
CanvasGroups need a shader applied just to work, so we have 2 shader
materials: the regular one (cg_material) and the outline one
(outline_material), and we swap as needed in the `highlight` and
`unhighlight` functions on Garden.

CanvasGroups are also immune to modulation for some reason even though
their contents are not, which means the CanvasModulate that simulates
day/night lighting does not affect the outline color.  to get around
this, we update the line_color param on `outline_material` within the
modulate (tint) node on every process tick (godot docs say that setting
shader params does not have a performance cost and this seems to be
accurate)