use discrete update mode for monster anims
there was an issue where the first frame of an anim was shorter than
subsequent frames, presumably because of lost time when handling
`_play_next` every loop.  changing the update mode to UPDATE_DISCRETE
(only update at the keyframes, according to godot's not-very-detailed
docs) seems to fix this, and is likely also more performant.

note: this works because monster sprite animations are sequences of
discrete states: we set the sprite's texture and metadata (hframes,
offset, etc) when changing animations, and simply update its `step`
property, which is an integer, on every subsequent frame.