action serialization
- minor refactor of code to update `monster.current_action` to fix a bug
  where actions were getting stuck in failure loops (presumably caused
  by `choose_action` calling `set_current_action` which caused the
  current action to fail if there already was one

- serialize actions! we do this with a recursive `save_keys` method on
  Action that calls a static `_save_keys` function for an action and
  all its superclasses (eg ApproachAction -> MoveAction -> Action).
  this is nice for not having to call `super.save_keys` like we do on
  subclasses of Entity

deserialization (ie the complicated part) is partially implemented.
we use a static `deserialize` method on Action that looks up the action
class from `name` on the serialized action, calls `new` on that and then
returns it.  we can customize this by overriding `_deserialize` on
subclasses, which we *need* to do for actions with required constructor
params (frequently target, sometimes dest/pos).

`Action.deserialize` then calls `save_keys` on the newly created action
and handles them the same way as in Entity & its subclasses, calling
`load_x` for key `x` if the function is implemented.

the part that's busted is that sometimes we need to do special loading
behavior on serialized properties that are passed into action subclass
constructors in `_deserialize` methods, ie target (a serialized target
is an object with uuid and a bunch of other fields we don't care about).
we still need to add a function to parse a serialized target and either
(a) call it in every `_deserialize` that passes a target into its
action's constructor or (b) change the way action constructors work.

probably we want to go with (b).  it's cool to make required params for
actions obvious from the constructor signature but it's not worth having
so much deserializtion boilerplate.  making all action constructors just
accept a monster and a serialized action would bring them to parity with
Entity, and allow us to do away with `_deserialize` entirely.

alternately, we implement static `parse_x` methods on Action that can be
called on matching serialized props in `deserialize`, before the input
is passed to `_deserialize`.