fix wander movement on load
the problem was actually in MoveAction - we were setting target position
(and target desired distance) on the monster's navigation agent in
`_start`, which isn't called when loading from state.  we now correctly
update these whenever we set the `dest` and `target_distance` properties
on the action.

this commit also refactors WanderAction to use a prerequisite MoveAction
instead of managing separate internal state for it. (wander dates back
to before the prerequisite system and was sort of a proof of concept for
it). this doesn't actually do anything to fix the bug, but it is a lot
nicer.