tweak main menu headers
replaces the style box used for the `title` & `page` elements in the
main menu with a plainer one that's just the color of the book
background, and makes them take up a tiny bit less vertical space.

i experimented with moving these to the bottom, or getting rid of them,
and the current layout actually looks best. this change makes the
headers fit in with the page a little bit better.