options & input
since selecting monsters is gonna mess with camera scrolling, we need to
nail down all the different scrolling behavior first.

previously we had the skeleton of a complicated system involving a
`ControlMode` enum that determined which scroll modes we could use.
in actuality, most aspects of the game (ie UI) should be able to handle
ALL types of input simultaneously (really there's just two: joypad/key,
which relies on navigating between elements with focus, and click/touch,
which requires every action to be achievable by clicking somewhere).

the exception is the cursor: that *will* have discrete modes depending
on what kind of input we last used (eg, touch input should put the
cursor in touch mode, which hides the cursor sprite and modifies
highlight/select behavior; joypad mode should work just like mouse mode
but make it stickier, etc). this is coming soon.

anyway, as for the camera, we should be able to handle all four types of
scroll (drag, edge, key, joystick) simultaneously since they all rely on
different input types. of these, only edge scroll can overlap with any
others, so we add a flag to enable or disable it.

this commit adds a new doc file `input.md` that contains a list of all
the actions we'll need to bind inputs to and what the default inputs
should be. (note that we can & should bind ui_accept, ui_cancel, and
the ui directional actions since those will actually be used to control
the ui, but we should make our own actions for everything else.)