docify function comments
godot 4.4 introduces tooltips for functions and variables that are
automatically populated with doc comments (## instead of #).
we historically have only ever used regular comments, so i had cursor
go through a bunch of files and double up the #s on function comments in
an attempt to take advantage of this New Editor Feature.

note that it still pretty much sucks because godot's doc formatter uses
bbcode rather than markdown for some ungodly reason. it doesn't even
support newlines (you have to use `[b]` or whatever) so all the
paragraphs get smooshed together.  i'm absolutely not gonna muck up my
comments with a bunch of dumb bbcode tags (i tried that once with
`validator.gd` and it wasn't worth it), so this is how it's gonna be
until they make it better. the extra context is still nice to have i
guess.

note note: apparently tooltips are not actually working for Validator
and its methods?? welp ¯\_(ツ)_/¯