fix default param bug in anim?
spawning a monster in the console errored here with "can't call `get` on
null". we're passing through a typeless input param from `monster` that
is null in this case - i guess that counts as explicitly passing null,
so it's not defaulted. edge case behavior!

`deserialize` actually expects a param so the safer behavior is to do
like in `monster.load_attributes` and set the input to `{}` if it's not
a dictionary.