"fix" camera
previously the camera was tied directly to mouse position. now we don't
even use mouse position since we're capturing the mouse and tracking
relative mouse movement to push the cursor around instead, so the camera
was totally broken.

basically we're gonna have to rewrite the entire camera node to properly
implement key, drag and edge scroll. this commit just comments out all
the junk we don't need and anchors the camera to the cursor for a quick
and dirty fix. and honestly it's not too bad, so i think we can punt on
fancy scroll modes for a while.